multi-instance-tenancy/app/pyproject.toml
[tool.poetry]
name = "kubernetes"
version = "0.1.0"
description = ""
authors = ["JulianSobott <julian.sobott@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.78.0"
uvicorn = "^0.17.6"
mysql-connector-python = "^8.0.31"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
